backend/requirements.txt
fastapi==0.115.5
numpy==2.1.3
Pillow==11.0.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.55.3
python-dotenv==1.0.1
passlib==1.7.4
PyJWT==2.9.0
